CreateTransactionService impl

+++ src/main/java/com/mendel/service_transaction/application/model/CreateTransactionCommand.java
@@ -1,6 +1,4 @@
 package com.mendel.service_transaction.application.model;
 
-import java.math.BigDecimal;
-
-public record CreateTransactionCommand(Long id, BigDecimal amount, String type, Long parentId) {
+public record CreateTransactionCommand(Long id, double amount, String type, Long parentId) {
 }
